Olog entry editor retake: refactoring attachments preview
