+++ app/logbook/olog/ui/src/main/resources/org/phoebus/logbook/olog/ui/AttachmentsPreview.fxml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<!--
+  ~ Copyright (C) 2020 European Spallation Source ERIC.
+  ~
+  ~  This program is free software; you can redistribute it and/or
+  ~  modify it under the terms of the GNU General Public License
+  ~  as published by the Free Software Foundation; either version 2
+  ~  of the License, or (at your option) any later version.
+  ~
+  ~  This program is distributed in the hope that it will be useful,
+  ~  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  ~  GNU General Public License for more details.
+  ~
+  ~  You should have received a copy of the GNU General Public License
+  ~  along with this program; if not, write to the Free Software
+  ~  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+  -->
+
+<?import javafx.scene.image.ImageView?>
+<?import javafx.geometry.Insets?>
+<SplitPane xmlns="http://javafx.com/javafx"
+           xmlns:fx="http://javafx.com/fxml"
+           fx:controller="org.phoebus.logbook.olog.ui.AttachmentsPreviewController"
+           fx:id="splitPane" dividerPositions="0.7">
+    <items>
+        <StackPane fx:id="previewPane">
+            <GridPane fx:id="noPreviewPane" visible="true" />
+            <ImageView fx:id="imagePreview" visible="false" />
+            <TextArea fx:id="textPreview" visible="false" />
+        </StackPane>
+        <VBox>
+            <children>
+                <ListView fx:id="attachmentListView">
+                    <placeholder>
+                        <Label text="No Attachments" />
+                    </placeholder>
+                </ListView>
+            </children>
+        </VBox>
+    </items>
+    <VBox.margin>
+        <Insets bottom="-10.0" left="-5.0" right="-5.0" top="-5.0" />
+    </VBox.margin>
+</SplitPane>
