+++ app/logbook/olog/ui/src/main/resources/org/phoebus/logbook/olog/ui/LogEntryDisplay.fxml
@@ -59,7 +59,7 @@
             <TitledPane fx:id="attachmentsPane" animated="false" text="Attachments">
               <content>
                   <AnchorPane minHeight="0.0" minWidth="0.0">
-                      <fx:include fx:id="attachments" source="LogAttachments.fxml" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
+                      <fx:include fx:id="attachments" source="write/AttachmentsView.fxml" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
                   </AnchorPane>
               </content>
             </TitledPane>
