+++ app/logbook/olog/ui/src/main/resources/org/phoebus/logbook/olog/ui/messages.properties
@@ -15,6 +15,7 @@ CurrentDate=Current Date
 Date=Date:
 EmbedImage=Embed Image
 EmbedImageDialogTitle=Add Embedded Image
+EmbedSelected=Embed Selected
 File=File
 ImageWidth=Width
 ImageHeight=Height
