+++ app/logbook/olog/ui/src/main/resources/org/phoebus/logbook/olog/ui/write/LogEntryEditor.fxml
@@ -19,20 +19,11 @@
   ~  
   -->
 
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ProgressIndicator?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.layout.VBox?>
+<?import javafx.geometry.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
 
-<?import javafx.scene.control.Tooltip?>
-<VBox fx:id="root" prefHeight="1000" prefWidth="800" HBox.hgrow="ALWAYS" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="org.phoebus.logbook.olog.ui.write.LogEntryEditorController">
+<VBox fx:id="root" prefHeight="1000" prefWidth="800" HBox.hgrow="ALWAYS" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.phoebus.logbook.olog.ui.write.LogEntryEditorController">
     <padding>
         <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
     </padding>
@@ -45,12 +36,16 @@
         </VBox.margin>
     </VBox>
 
-    <VBox fx:id="attachments" alignment="CENTER" VBox.vgrow="NEVER">
-        <fx:include fx:id="attachmentsView" source="AttachmentsView.fxml" />
-    </VBox>
-    <VBox fx:id="properties" alignment="CENTER" VBox.vgrow="NEVER">
-        <fx:include fx:id="propertiesView" source="LogPropertiesEditor.fxml" />
-    </VBox>
+    <Accordion maxHeight="1.7976931348623157E308" VBox.vgrow="ALWAYS">
+        <panes>
+            <TitledPane animated="false" maxHeight="1.7976931348623157E308" text="%Attachments">
+                <fx:include fx:id="attachmentsView" maxHeight="1.7976931348623157E308" source="AttachmentsView.fxml" />
+            </TitledPane>
+            <TitledPane animated="false" text="%Properties">
+                <fx:include fx:id="propertiesView" source="LogPropertiesEditor.fxml" />
+            </TitledPane>
+        </panes>
+    </Accordion>
 
     <HBox alignment="BOTTOM_CENTER" fillHeight="false">
         <children>
@@ -66,19 +61,17 @@
             </rowConstraints>
             </GridPane>
             <Pane HBox.hgrow="ALWAYS" />
-            <Button fx:id="cancel" alignment="CENTER_RIGHT" minWidth="57.0" mnemonicParsing="false"
-                    onAction="#cancel" text="%Cancel">
+            <Button fx:id="cancel" alignment="CENTER_RIGHT" minWidth="57.0" mnemonicParsing="false" onAction="#cancel" text="%Cancel">
                 <tooltip>
-                    <Tooltip text="%CancelTooltip"/>
+                    <Tooltip text="%CancelTooltip" />
                 </tooltip>
                 <HBox.margin>
                     <Insets left="5.0" right="5.0" />
                 </HBox.margin>
             </Button>
-            <Button fx:id="submit" alignment="CENTER_RIGHT" minWidth="57.0" mnemonicParsing="false"
-                    onAction="#submit" text="%Submit">
+            <Button fx:id="submit" alignment="CENTER_RIGHT" minWidth="57.0" mnemonicParsing="false" onAction="#submit" text="%Submit">
                 <tooltip>
-                    <Tooltip text="%SubmitTooltip"/>
+                    <Tooltip text="%SubmitTooltip" />
                 </tooltip>
                 <HBox.margin>
                     <Insets left="5.0" />
