+++ app/logbook/olog/ui/src/main/resources/org/phoebus/logbook/olog/ui/write/LogPropertiesEditor.fxml
@@ -16,58 +16,50 @@
 <?import javafx.scene.control.Label?>
 <VBox xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.phoebus.logbook.olog.ui.write.LogPropertiesEditorController">
    <children>
-      <Accordion>
-        <panes>
-          <TitledPane animated="false" text="Properties">
-               <content>
-                  <GridPane minHeight="200.0" minWidth="450.0">
-                    <columnConstraints>
-                      <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                      <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                    </columnConstraints>
-                    <rowConstraints>
-                      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                    </rowConstraints>
-                     <children>
-                        <TreeTableView fx:id="selectedPropertiesTree" editable="true" GridPane.hgrow="ALWAYS" GridPane.rowSpan="2" GridPane.vgrow="ALWAYS">
-                            <placeholder>
-                                <Label text=""/>
-                            </placeholder>
-                            <columns>
-                            <TreeTableColumn fx:id="name" editable="false" prefWidth="75.0" text="Name" />
-                            <TreeTableColumn fx:id="value" prefWidth="75.0" text="Value" />
-                          </columns>
-                           <columnResizePolicy>
-                              <TreeTableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-                           </columnResizePolicy>
-                        </TreeTableView>
-                        <TableView fx:id="availablePropertiesView" GridPane.columnIndex="1" GridPane.vgrow="ALWAYS">
-                          <placeholder>
-                              <Label text=""/>
-                          </placeholder>
-                          <columns>
-                            <TableColumn fx:id="propertyName" onEditStart="#availablePropertySelection" prefWidth="75.0" text="Available Properties" />
-                          </columns>
-                           <columnResizePolicy>
-                              <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-                           </columnResizePolicy>
-                           <GridPane.margin>
-                              <Insets left="2.0" />
-                           </GridPane.margin>
-                        </TableView>
-                        <Button mnemonicParsing="false" onMouseClicked="#availablePropertySelection" prefHeight="25.0" text="Add Selected Property" GridPane.columnIndex="1" GridPane.rowIndex="1">
-                           <GridPane.margin>
-                              <Insets left="2.0" />
-                           </GridPane.margin></Button>
-                     </children>
-                     <padding>
-                        <Insets bottom="2.0" left="2.0" right="2.0" top="2.0" />
-                     </padding>
-                  </GridPane>
-               </content>
-            </TitledPane>
-        </panes>
-      </Accordion>
+      <GridPane minHeight="200.0" minWidth="450.0">
+        <columnConstraints>
+          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+        </columnConstraints>
+        <rowConstraints>
+          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+        </rowConstraints>
+         <children>
+            <TreeTableView fx:id="selectedPropertiesTree" editable="true" GridPane.hgrow="ALWAYS" GridPane.rowSpan="2" GridPane.vgrow="ALWAYS">
+                <placeholder>
+                    <Label text=""/>
+                </placeholder>
+                <columns>
+                <TreeTableColumn fx:id="name" editable="false" prefWidth="75.0" text="Name" />
+                <TreeTableColumn fx:id="value" prefWidth="75.0" text="Value" />
+              </columns>
+               <columnResizePolicy>
+                  <TreeTableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+               </columnResizePolicy>
+            </TreeTableView>
+            <TableView fx:id="availablePropertiesView" GridPane.columnIndex="1" GridPane.vgrow="ALWAYS">
+              <placeholder>
+                  <Label text=""/>
+              </placeholder>
+              <columns>
+                <TableColumn fx:id="propertyName" onEditStart="#availablePropertySelection" prefWidth="75.0" text="Available Properties" />
+              </columns>
+               <columnResizePolicy>
+                  <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+               </columnResizePolicy>
+               <GridPane.margin>
+                  <Insets left="2.0" />
+               </GridPane.margin>
+            </TableView>
+            <Button mnemonicParsing="false" onMouseClicked="#availablePropertySelection" prefHeight="25.0" text="Add Selected Property" GridPane.columnIndex="1" GridPane.rowIndex="1">
+               <GridPane.margin>
+                  <Insets left="2.0" />
+               </GridPane.margin></Button>
+         </children>
+         <padding>
+            <Insets bottom="2.0" left="2.0" right="2.0" top="2.0" />
+         </padding>
+      </GridPane>
    </children>
 </VBox>
